+++ docs/changelog.rst
@@ -74,6 +74,12 @@ Documentation
 - Clarified ``skip_diff`` documentation (#1655)
 - Improved documentation relating to validation on import (#1665)
 
+3.3.5 (unreleased)
+------------------
+
+- Nothing changed yet.
+- Remove unnecessary ChangeList queries to speed up export via Admin UI (#1715)
+
 3.3.4 (2023-12-09)
 ------------------